🌊 Feat: 파도타기 화면 구현 완료
파도타기 새 디자인에 맞춰 로직 구현
