+++ PADO/PADO/Assets.xcassets/TabBarImage/tab_added.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "tab_added.svg",
+      "filename" : "tap_added.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
