+++ PADO/PADO/iOS/PadoRide/ViewModel/PadoRideViewModel.swift
@@ -139,12 +139,10 @@ class PadoRideViewModel: ObservableObject {
     }
     
     // 특정 ID들에 대한 포스트 데이터를 미리 로드
-    func preloadPostsData(for ids: [String]) async {
-        for id in ids {
-            let posts = await getPostData.suffingPostData(id: id)
-            DispatchQueue.main.async {
-                self.postsData[id] = posts
-            }
+    func loadPostsData(for id: String) async {
+        let posts = await getPostData.suffingPostData(id: id)
+        DispatchQueue.main.async {
+            self.postsData[id] = posts
         }
     }
